Refactor Search Facade to their respective area with tests
