+++ app/controllers/api/v1/image_controller.rb
@@ -1,6 +1,6 @@
 class Api::V1::ImageController < ApplicationController
   def index
-    image = SearchFacade.get_image(params[:location])
+    image = ImageFacade.get_image(params[:location])
     render json: ImageSerializer.new(image)
   end
 end
