+++ app/controllers/api/v1/weather_controller.rb
@@ -1,6 +1,6 @@
 class Api::V1::WeatherController < ApplicationController
   def index
-    forcast = SearchFacade.forcast(params[:location])
+    forcast = WeatherFacade.forcast(params[:location])
     render json: ForecastSerializer.new(forcast)
   end
 end
