+++ app/facades/image_facade.rb
@@ -1,0 +1,6 @@
+class ImageFacade
+  def self.get_image(location)
+    data = ImageService.get_image(location)
+    Image.new(data, location)
+  end
+end
